requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
databases[sqlite]==0.9.0
sqlalchemy==2.0.29
pydantic==2.7.1
python-multipart==0.0.9
itsdangerous==2.1.2
websockets>=11.0.3,<13.0
aiofiles==23.2.1
httpx==0.27.0   